Added code and battle image in fodler
Attempted to make background image the battlescene but not work.

+++ game_scene.cpp
@@ -79,8 +79,8 @@ void Game_Scene::update(SDL_Window* window)
 	const int border = 350;
 	const int lower_border = 1600-450;
 	int w, h;
-	int movex = player->translation().x();
-	int movey = player->translation().y();
+	float movex = player->translation().x();
+	float movey = player->translation().y();
 	movex = player->translation().x();
 	movey = player->translation().y();
 	//std::cout << movex;
